Store third-party in separate directory (#812)
