+++ tests/cookie_test_2.cc
@@ -1,7 +1,7 @@
 #include "gtest/gtest.h"
 
 #include <pistache/cookie.h>
-#include <pistache/date.h>
+#include <pistache/thirdparty/date.h>
 
 using namespace Pistache;
 using namespace Pistache::Http;
