+++ tests/headers_test.cc
@@ -1,5 +1,5 @@
-#include <pistache/date.h>
 #include <pistache/http.h>
+#include <pistache/thirdparty/date.h>
 
 #include "gtest/gtest.h"
